Add Offline map download to other activities extending BaseActivity
1. Refactor BaseRegisterActivity & BaseActivity
