+++ opensrp-path/src/main/java/org/smartregister/path/map/DisplayToastInterface.java
@@ -1,0 +1,10 @@
+package org.smartregister.path.map;
+
+/**
+ * Created by Ephraim Kigamba - ekigamba@ona.io on 02/02/2018.
+ */
+
+public interface DisplayToastInterface {
+
+    void showInfoToast(String text);
+}
